Moved Categories Block edit to an external file.
An intermediary step to use withAPIData.
